商品の登録更新機能

+++ html/list.php
@@ -2,16 +2,19 @@
   require "./inc/db.php";
   $query = pg_query($dbconnect,"SELECT id, name, price FROM products");
   $results = pg_fetch_all($query);
-  $deleteId = $_GET['id'];
-  $daleteQuery = pg_query($dbconnect, "DELETE FROM products WHERE id = '$deleteId'");
-  if ($daleteQuery) {
-    header('Location: ./list.php');
-    exit;
+  $deleteId = (int)$_GET['id'];
+  if ($deleteId) {
+    $daleteQuery = pg_query($dbconnect, "DELETE FROM products WHERE id = '$deleteId'");
+    if ($daleteQuery) {
+      header('Location: ./list.php');
+      exit;
+    }
   }
   foreach ($results as $key => $val) {
     $html .= '<tr>';
     $html .= '<td>' . $val["name"]. '</td>';
     $html .= '<td>' . $val["price"]. '</td>';
+    $html .= '<td><a href="./form.php?id=' . $val["id"] . '">変更</a></td>';
     $html .= '<td><a href="./list.php?id=' . $val["id"] . '">削除</a></td>';
     $html .= '</tr>';
   }
@@ -29,6 +32,7 @@
     <tr>
       <th>商品名</th>
       <th>価格</th>
+      <th>変更</th>
       <th>削除</th>
     </tr>   
     <?= $html ?>
